Make InputDataInjector parametrisable
This commit will allow the user of this library to specify under which
property in the deserialisation context should the generated id be
added.
This will allow an easier use of Chimera's CreateAndFetch annotation,
since the command and the query will be able to align on the same
property without having to rely Chimera's old _input.id property.

To remain backwards compatible, the InputDataInjector now defaults to
the old configuration should a new one not be provided.